moved user routes, added jwt token install
