+++ controller/spaceRoutes.js
@@ -1,0 +1,8 @@
+import { Router } from 'express'
+// IMPORT NEEDED MODELS
+// import { Space } from ''
+// import { Area } from ''
+
+// Create a new instance of Router object:
+const router = Router()
+
